requirements.txt
langchain==0.3.27
ragas
langchain-openai
langchain-community
langchain-qdrant
qdrant-client
pypdf
flashrank
python-dotenv
langgraph
streamlit
pytest 
pytest-mock
requests
